--- ms2deepscore/benchmarking/__init__.py
@@ -1,5 +1,0 @@
-from .helper_functions import bin_dependent_losses
-
-__all__ = [
-    "bin_dependent_losses",
-]
