+++ ms2deepscore/models/loss_functions.py
@@ -1,0 +1,117 @@
+import numpy as np
+import torch
+from torch import nn
+
+
+def rmse_loss(outputs, targets):
+    return torch.sqrt(torch.mean((outputs - targets) ** 2))
+
+
+def risk_aware_mae(outputs, targets):
+    """MAE weighted by target position on scale 0 to 1.
+    """
+    factors = targets  # this is meant for a uniform distribution of targets between 0 and 1.
+
+    errors = targets - outputs
+    uppers =  factors * errors
+    lowers = (factors - 1) * errors
+
+    losses = torch.max(lowers, uppers)
+    return losses.mean()
+
+
+def risk_aware_mse(outputs, targets):
+    """MSE weighted by target position on scale 0 to 1.
+    """
+    factors = targets  # this is meant for a uniform distribution of targets between 0 and 1.
+
+    errors = targets - outputs
+    errors = torch.sign(errors) * errors ** 2
+    uppers =  factors * errors
+    lowers = (factors - 1) * errors
+
+    losses = torch.max(lowers, uppers)
+    return losses.mean()
+
+
+class RiskAwareMAE(nn.Module):
+    """Loss functions taking into account the actual distribution of the target labels"""
+    def __init__(self, percentiles=None, device="cpu"):
+        super().__init__()
+        self.device = device
+        if percentiles is None:
+            self.percentiles = torch.linspace(0.01, 1.0, 100)
+        else:
+            self.percentiles = percentiles
+
+    def forward(self, outputs, targets):
+        device = self.device
+        idx = torch.empty((len(targets)))
+        for i, target in enumerate(targets):
+            idx[i] = torch.argmin(torch.abs(self.percentiles.to(device) - target.to(device)))
+
+        max_bin = self.percentiles.shape[0]
+        factors = (idx + 1) / max_bin
+
+        errors = targets.to(device) - outputs.to(device)
+        uppers =  factors.to(device) * errors
+        lowers = (factors.to(device) - 1) * errors
+
+        losses = torch.max(lowers, uppers)
+        return losses.mean()
+
+
+LOSS_FUNCTIONS = {
+    "mse": nn.MSELoss(),
+    "rmse": rmse_loss,
+    "risk_mae": risk_aware_mae,
+    "risk_mse": risk_aware_mse,
+}
+
+#todo integrate with loss functions in model helper functions
+def bin_dependent_losses(predictions,
+                         true_values,
+                         ref_score_bins,
+                         loss_types=("mse",),
+                         ):
+    """Compute errors (RMSE and MSE) for different bins of the reference scores (scores_ref).
+
+    Parameters
+    ----------
+    predictions
+        Scores that should be evaluated
+    true_values
+        Reference scores (= ground truth).
+    ref_score_bins
+        Bins for the refernce score to evaluate the performance of scores.
+    loss_types
+        Specify list of loss types out of "mse", "mae", "rmse".
+    """
+    bin_content = []
+    losses = {
+        "bin": [],
+        "mae": [],
+        "mse": [],
+        "rmse": [],
+    }
+    # maes = []
+    bounds = []
+    ref_scores_bins_inclusive = ref_score_bins.copy()
+    for i in range(len(ref_scores_bins_inclusive) - 1):
+        low = ref_scores_bins_inclusive[i]
+        high = ref_scores_bins_inclusive[i + 1]
+        bounds.append((low, high))
+        idx = np.where((true_values >= low) & (true_values < high))
+        bin_content.append(idx[0].shape[0])
+        # Add values
+        losses["bin"].append((low, high))
+        if "mae" in loss_types:
+            mae = np.abs(true_values[idx] - predictions[idx]).mean()
+            losses["mae"].append(mae)
+        if "mse" in loss_types:
+            mse = np.square(true_values[idx] - predictions[idx]).mean()
+            losses["mse"].append(mse)
+        if "rmse" in loss_types:
+            rmse = np.sqrt(np.square(true_values[idx] - predictions[idx]).mean())
+            losses["rmse"].append(rmse)
+    return bin_content, bounds, losses
